+++ libs/server/gamification/src/lib/core/events/index.ts
@@ -1,0 +1,2 @@
+export * from './unified-event-handler.service';
+export * from './event-types';
