+++ libs/server/gamification/src/lib/core/services/index.ts
@@ -1,0 +1,3 @@
+export * from './badge.service';
+export * from './level.service';
+export * from './reward.service';
